requirements.txt
PyQt6>=6.4.0
python-vlc>=3.0.18122
mutagen>=1.46.0
numpy>=1.24.0
requests